requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
idna==3.10
imap-tools==1.7.3
ipykernel==6.29.4
ipython==8.23.0
is-disposable-email==1.0.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.6.0
keyboard==0.13.5
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
mysql-connector-python==9.0.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.8
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyodbc==5.1.0
pyparsing==3.1.4
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
pywin32==308
pyzmq==25.1.2
qrcode==8.0
requests==2.32.3
rich==13.9.2
seaborn==0.13.2
selenium==4.25.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
tabula-py==2.9.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
wtforms-validators==1.0.0
xmltodict==0.14.1
